Refactor: Rename the time_t to String function to toString - Move it to BsString and fixed code that used it.
